wip: file structure changes -- working on panes
