+++ src/components/BillSummary/BillSummary.stories.tsx
@@ -3,7 +3,7 @@ import Container from '@material-ui/core/Container';
 
 import Component, { Bill } from './BillSummary';
 
-export default { title: 'Domains/Townhall/Townhall Bill Summary' };
+export default { title: 'Components/Bill Summary' };
 
 const bill: Bill = {
     name: 'Darth Vader',
