+++ src/domains/Pane/PaneNavigation/index.ts
@@ -1,0 +1,1 @@
+export { default } from './PaneNavigation';
