+++ src/domains/Speaker/SpeakerHistory/SpeakerHistory.test.tsx
@@ -2,33 +2,32 @@ import React from 'react';
 import { render, unmountComponentAtNode } from 'react-dom';
 import ReactTestUtils from 'react-dom/test-utils';
 
-import Component, { PieDatum } from './TownhallNivo';
+import Component, { TownHallHistoryEntry } from './SpeakerHistory';
 
-const Nivo: PieDatum[] = [
+const entries: TownHallHistoryEntry[] = [
     {
-        id: 'php',
-        value: 344,
+        action: 'Event 1',
+        link: 'https://i.imgur.com/3beQH5s.jpeg',
+        date: 'Jan 9, 2014',
     },
     {
-        id: 'go',
-        value: 262,
+        action: 'Event 2',
+        link: 'https://i.imgur.com/3beQH5s.jpeg',
+        date: 'Jan 10, 2015',
     },
     {
-        id: 'javascript',
-        value: 411,
+        action: 'Event 3',
+        link: 'https://i.imgur.com/3beQH5s.jpeg',
+        date: 'Jan 11, 2016',
     },
     {
-        id: 'ruby',
-        value: 191,
-    },
-    {
-        id: 'hack',
-        value: 234,
-   
+        action: 'Event 4',
+        link: 'https://i.imgur.com/3beQH5s.jpeg',
+        date: 'Jan 12, 2017',
     },
 ];
 
-describe('TownHallNivo', function () {
+describe('Speaker History', function () {
     let container: HTMLDivElement | null = null;
 
     beforeEach(() => {
@@ -48,7 +47,7 @@ describe('TownHallNivo', function () {
     // eslint-disable-next-line jest/expect-expect
     it('should render', () => {
         ReactTestUtils.act(() => {
-            render(<Component data={Nivo} height={200} />, container);
+            render(<Component history={entries} />, container);
         });
     });
 });
