+++ src/domains/Speaker/SpeakerHistory/index.ts
@@ -1,0 +1,1 @@
+export { default } from './SpeakerHistory';
