--- src/domains/Townhall/JoinUrl/JoinUrl.stories.tsx
@@ -1,12 +1,0 @@
-import React from 'react';
-import Component from './JoinUrl';
-
-export default { title: 'Domains/Townhall/JoinUrl' };
-
-export function Basic() {
-    return (
-        <div style={{ maxWidth: '1000px' }}>
-            <Component />
-        </div>
-    );
-}
