--- src/domains/Townhall/TownHallHistory/index.ts
@@ -1,1 +1,0 @@
-export { default } from './TownHallHistory';
