+++ src/domains/Townhall/TownhallCard/index.ts
@@ -1,0 +1,1 @@
+export { default } from './TownhallCard';
