--- src/domains/Townhall/TownhallNivo/index.ts
@@ -1,1 +1,0 @@
-export { default } from './TownhallNivo';
