+++ src/domains/Townhall/TownhallPane/TownhallPane.stories.tsx
@@ -2,9 +2,9 @@ import React from 'react';
 
 import Page from 'layout/Page';
 import TownhallProvider from '../Contexts/Townhall';
-import Component from './Information';
+import Component from './TownhallPane';
 
-export default { title: 'Domains/Townhall/Information' };
+export default { title: 'Domains/Townhall/Townhall Pane' };
 
 export function Basic() {
     return (
